cleaned code after merge

+++ oscar-cbls/src/main/examples/oscar/examples/cbls/geometry/TestGeometryPacking.scala
@@ -188,7 +188,7 @@ object TestGeometryPacking extends App{
     () => {
       val allShapes = placedShapes.map(_.value)
 
-      val holes:Iterable[(Int,Int)] = Overlap.centroidsOfFreeSpacesIn(allShapes.map(_.geometry),outerFrame)
+      val holes:Iterable[(Long,Long)] = Overlap.centroidsOfFreeSpacesIn(allShapes.map(_.geometry),outerFrame)
 
       allShapes.indices.flatMap(shapeID => {
